requirements.txt
assemblyai==0.26.0
datasets==2.19.2
jiwer==3.0.4
openai==1.30.1
pandas==2.2.2
pydub==0.25.1
whisper_normalizer==0.0.8
speechmatics-python==1.8.3
python-dotenv==1.0.0
tabulate==0.9.0